Fix distance threshold comparison in _is_success method and update gym environment in test script

+++ mycobotgym/envs/mycobot.py
@@ -284,7 +284,7 @@ class MyCobotEnv(MujocoEnv):
 
     def _is_success(self, achieved_goal, desired_goal):
         d = goal_distance(achieved_goal, desired_goal)
-        return (d < self.distance_threshold).astype(np.float32)
+        return (d < self.distance_threshold)
 
     def compute_reward(self, achieved_goal, goal, info):
         # Compute distance between goal and the achieved goal.
